added cross validation mode, support parameters estimation through cv

+++ ml/src/main/java/com/spbsu/ml/cli/modes/impl/Fit.java
@@ -50,8 +50,8 @@ public class Fit extends AbstractMode {
     if (command.hasOption(CROSS_VALIDATION_OPTION)) {
       final DataBuilderCrossValidation dataBuilderCrossValidation = new DataBuilderCrossValidation();
       final String[] cvOptions = StringUtils.split(command.getOptionValue(CROSS_VALIDATION_OPTION), "/", 2);
-      dataBuilderCrossValidation.setRandomSeed(Integer.valueOf(cvOptions[0]));
-      dataBuilderCrossValidation.setPartition(Double.valueOf(cvOptions[1]));
+      dataBuilderCrossValidation.setRandomSeed(Long.valueOf(cvOptions[0]));
+      dataBuilderCrossValidation.setPartition(cvOptions[1]);
       dataBuilder = dataBuilderCrossValidation;
     } else {
       dataBuilder = new DataBuilderClassic();
